fix(HappyLearning): change totaltime from Integer to Float
Allow fractional seconds for courseware total duration.

+++ OfficalPlugin/SwissKitJ-Plugin-HappyLearning/src/main/java/plugin/swisskit/hpl/dto/UserLearnCourseWareVOList.java
@@ -20,7 +20,7 @@ public class UserLearnCourseWareVOList {
     @JSONField(name = "learnprogress")
     private Float learnProgress;
 
-    private Integer totaltime;
+    private Float totaltime;
     private Integer passed;
     private Integer status;
     private Integer isallowlearn;
